requirements.txt
aiohttp == 3.8.1
aiohttp-auth-autz == 0.2.2
sqlalchemy == 1.4.40
asyncpg == 0.26.0
pyyaml == 6.0
alembic == 1.8.1

